+++ packages/@unimodules/core/ios/UMCore/UMModuleRegistryProvider/UMModuleRegistryProvider.h
@@ -1,7 +1,0 @@
-// Copyright © 2018 650 Industries. All rights reserved.
-
-#import <ExpoModulesCore/EXUnimodulesCompat.h>
-#import <ExpoModulesCore/EXModuleRegistryProvider.h>
-#import <ExpoModulesCore/EXModuleRegistry.h>
-
-typedef UM_DEPRECATED(ModuleRegistryProvider) EXModuleRegistryProvider UMModuleRegistryProvider;