+++ packages/expo-modules-core/src/NativeViewManagerAdapter.tsx
@@ -3,5 +3,5 @@ import React from 'react';
 import { UnavailabilityError } from './errors/UnavailabilityError';
 
 export function requireNativeViewManager<P = any>(viewName: string): React.ComponentType<P> {
-  throw new UnavailabilityError('@unimodules/react-native-adapter', 'requireNativeViewManager');
+  throw new UnavailabilityError('expo-modules-core', 'requireNativeViewManager');
 }